[style/#212] 게임 헤더 컴포넌트 분리 작업
